Execute unit tests in parallel (#331)
* Wrap test estimator checks in a python unit test.

* Combine the non-extended test runs together to make them more parallelizable.

* Reverse the tests path args order to try and have test_estimator_checks run earlier in the test run.

* Dynamically generate the test_estimator_checks unit tests.

* Create the test_docs_example unit tests dynamically so they can be parallelized.

* Fix KMeansPlusPlus does not work with a cluster size of 1 when using a debug version of ml.net

* Fix OLS divide by 0 when given a particular set of inputs to fit. This is hidden in release versions of ml.net

* Fix issue when ranking where the output of TextToKeyConverter was
trying to overwrite the $scoredVectorData variable set by
DatasetScorerEx. See test_metrics_evaluate_ranking_group_id_from_existing_column_in_X
for a test which demonstrates the issue. It throws an exception
from EntryPointNode.cs:837 when trying to get the outputs. The
exception was hidden when using release builds of ML.Net.

* Remove a test_estimator_check for OrdinaryLeastSquaresRegressor
since it is causing invalid float values and throwing an exception
which was hidden in release versions of ML.Net but visible in debug.

* Update test_permutation_feature_importance tests to support parallel execution.

* Rerun unit tests one extra time if any failed to check for intermittent failures.